https://github.com/openNAMU/openNAMU/issues/2070

+++ route/edit_request.py
@@ -41,27 +41,45 @@ def edit_request(name = 'Test', do_type = ''):
             if acl_check(name, 'document_edit') == 1:
                 return redirect('/w/' + url_pas(name))
             
+            curs.execute(db_change("select data from data where title = ?"), [name])
+            db_data = curs.fetchall()
+            o_data = db_data[0][0] if db_data else ''
+            
             curs.execute(db_change("select user from scan where title = ? and type = ''"), [name])
             for scan_user in curs.fetchall():
                 add_alarm(scan_user[0], edit_request_user, '<a href="/w/' + url_pas(name) + '">' + html.escape(name) + '</a>')
 
-            curs.execute(db_change("delete from data where title = ?"), [name])
-            curs.execute(db_change("insert into data (title, data) values (?, ?)"), [name, edit_request_data])
-                    
-            history_plus(
-                name,
-                edit_request_data,
-                edit_request_date,
-                edit_request_user,
-                edit_request_send,
-                edit_request_leng
-            )
-            
-            render_set(
-                doc_name = name,
-                doc_data = edit_request_data,
-                data_type = 'backlink'
-            )
+            if flask.request.form.get('check', '') == 'Y':
+                curs.execute(db_change("delete from data where title = ?"), [name])
+                curs.execute(db_change("insert into data (title, data) values (?, ?)"), [name, edit_request_data])
+                        
+                history_plus(
+                    name,
+                    edit_request_data,
+                    edit_request_date,
+                    edit_request_user,
+                    edit_request_send,
+                    edit_request_leng,
+                    t_check = 'approve',
+                    mode = 'edit_request'
+                )
+                
+                render_set(
+                    doc_name = name,
+                    doc_data = edit_request_data,
+                    data_type = 'backlink'
+                )
+            else:
+                history_plus(
+                    name,
+                    o_data,
+                    edit_request_date,
+                    edit_request_user,
+                    edit_request_send,
+                    '0',
+                    t_check = 'decline',
+                    mode = 'edit_request'
+                )
                 
             if do_type == 'from':
                 return redirect('/edit/' + url_pas(name))
@@ -87,7 +105,7 @@ def edit_request(name = 'Test', do_type = ''):
                     <hr class="main_hr">
                     <form method="post">
                         <button id="opennamu_save_button" type="submit" name="check" value="Y">''' + load_lang('approve') + '''</button>
-                        <button id="opennamu_preview_button" type="button" name="check" value="">''' + load_lang('decline') + '''</button>
+                        <button id="opennamu_preview_button" type="submit" name="check" value="">''' + load_lang('decline') + '''</button>
                         <hr class="main_hr">
                         <textarea readonly class="opennamu_textarea_500">''' + html.escape(edit_request_data) + '''</textarea>
                     </form>
